Back out "[fboss] Decrease LACP logging verbosity"
Summary:
We had originally decreased LACP logging verbosity to avoid triggering
blocking logging statements, which were observed in a period of high route churn
in production. The logging in this period was both bursty and significantly more
voluminious than that of LACP. Moreover, LACP logging had been enabled in
production without inducing a stall for months.

For these reasons, we believe LACP logging is safe to re-enable.

Reviewed By: boryas

Differential Revision: D7050251

fbshipit-source-id: 6d47b22c2fc23e3e13be10440dbd86ac54a077a4