Accomodate for missing FP groups when deciding whether things need to be
Summary:
On warm boot, we look for groups that need to be (re)setup since
their QSETs changed. However we don't account for groups that need to be
newly added. This actually is a bug in our existing code as well. But with
the QSET changes it got exposed since we look for QSET changes on a
unconfigured group.

To fix revise the approach to always setup missing groups and look for
QSET changes for existing groups.

TODO: bcm_test/fake_sdk_test for fpGroupExists API

Reviewed By: SamanKazemkhani

Differential Revision: D7064200

fbshipit-source-id: 77376a12f51d47e7bc1d9b6c9e41b9dcc379b9cd