Reinit FP on warm boot till T26329911 is solved
Summary:
Reinit FP on warm boots till T26329911 is solved.
The problem seen there is that probing FP groups for comparison
during warm boot messes up the existing FP groups on wedge100.
The strange thing here is that FP groups are identical before
and after when compared via fp show on bcm shell.

I am following up with BRCM on this. But in the meanwhile
go back to earlier state of reiniting FP till we figure this
out. This lets us move forward on rolling out DST mac changes
to ACL.
This is relatively safe, consider
- Outside of FA only thing using acls is for CPU traffic. During
warm boot the (re)setup of FP groups happens during init. Which
is before we have setup TunManager or the packet rx call back.
So there is nothing to congest the CPU pipeline yet.
- For FAs we are still doing cold boots. And even when we goto
warm boots on FAs, we would still have the box drained during
agent restart. FP would be setup before undrain.
- We used to reinit the FP always till we added warm boot for
ACLs a couple of months ago.

This is not to say that the above behavior does not need to be
debugged. But b/w just holding off on rolling DST mac into ACLs
and reiniting the FP, latter seems to be a better way forward.

Reviewed By: boryas

Differential Revision: D7078032

fbshipit-source-id: f4ae091a28c0ac32a3352a238eb4b59ae24f0829