detect nic vendors by mac, not firmware
Summary:
The mechanism we use to detect nic vendors currently does not
work during provisioning, which is causing us to start switches w/out
fec in a dangerous way. This changes the serf querying logic s.t it
will work during provisioning--by looking at the ouis of the mac
addresses of serf servers in our racks.

Differential Revision: D7130525

fbshipit-source-id: 08f18fc3a15576dccb8317290585c7e2395c5f68