Lift EventBase scheduling out of LACP logic
Summary:
Previously, each start()/stop() method on an LACP machine scheduled its
work on the LACP EventBase.

This scheduling of work is lifted out of the callsite of the start()/stop()
methods and into LacpController::{start,stop}Machines(), which makes it
consistent with the location of other EventBase scheduling calls.

Reviewed By: boryas

Differential Revision: D7227781

fbshipit-source-id: a9d5fc8b9ab4e87c15974d6aa12bcb9f27b24414