Serialize shared_ptr construction with {start,stop}Machines()
Summary:
A pre-condition of invoking shared_from_this() is that the "this"
object be already managed by std::shared_ptr.

This condition could previously be violated by the update thread, in which an
LacpController is made to be managed by a shared_ptr, and the LACP thread, which
invokes shared_from_this on an LacpController.

To illustrate this, assume we have configured an AggregatePort A with a single
member port P and consider the point at which
LinkAggregationManager::stateUpdate(delta) is observing the transition to the
SwitchState derived from the config. At this point, initialized_ is true, a
LacpController C corresponds to P, and
LinkAggregationManager::aggregatePortAdded(A) is invoked.

The following execution (sub)sequence was previously possible:

| |Update Thread                  | LACP thread
| |-------------------------------|------------
|1|new LacpController (call it C')|-
|2|C'::LacpController(...)        |-
|3|C'::startMachines()            |-
|4|C'::periodicTx_::start()       |-
|5|-                              |C'::periodicTx::determineTransmissionRate()
|6|-                              |C'::actorInfo()
|7|-                              |shared_from_this()

Because shared_from_this() occurs before
std::shared_ptr<LacpController>::reset(new LacpController(...)) has wrapped C'
in a shared_ptr, a std::bad_weak_ptr exception is thrown.

To avoid this, the following modifications are made:

-LacpController::startMachines() occurs after LacpController is managed by a
 shared_ptr
-LacpController::stopMachines() occurs before LacpController is no longer
 managed by a shared_ptr

An added benefit is that LacpController's constructor and destructor now only
perform trivial work, which I think is a C++ best practice.

Depends on D7227781

Reviewed By: boryas

Differential Revision: D7227785

fbshipit-source-id: 2e2680d955b6aa2a7873374ca43a0ee4e2863407