Removed default AdminDistance in RouteNextHopEntry
Summary:
Refactored `RouteNextHopEntry` constructors to no longer receive optional `AdminDistance` values. Client code must now be explicit about the `AdminDistance` required.

I did provide `AdminDistance::MAX_ADMIN_DISTANCE` as the explicit value in all the code that broke as a result of this change. It was mostly in tests that I don't believe used the value anyway.

The `RouteNextHopEntry(Action, AdminDistance)` constructor broke a few more things: (1) its use as a value type in a `flat_map` requiring a parameterless constructor and (2) its use in `fromFollyDynamic` which now uses it with `Action::DROP` before replacing the `action_` with whatever is in the `dynamic` value (to avoid the `CHECK_NE`).

Reviewed By: tfangit

Differential Revision: D7191680

fbshipit-source-id: 158fe00a1c97f4dbfea4dbe8695f0aba89925c6a