Added new OUIs
Summary:
We were missing OUI's for servers appearing that
are not in the list. This is causing coop to crash because
it tries too lookup OUIs for all server nics to see if they
are mellanox (so it can enable FEC)

Created with this horrible perl one-liner after removing the contents
of the existing dict.

http://standards-oui.ieee.org/oui.txt /tmp/oui.txt
  cat /tmp//oui.txt | perl -e "while (<STDIN>){ s/[\r\n]//g; s/[^A-Za-z0-9 \.\,\-]+//g; if (/^[0-9A-F]{2}-[0-9A-F]{2}-[0-9A-F]{2}/) { (\$oui, \$b, \$inc) = split(/\s+/, \$_, 3); \$oui =~ s/\-/:/g; \$inc =~ s/'//g; print (\"\'\$oui\':\'\$inc\',\n\"); }}"  | sort -t: -k 4

Reviewed By: Natolumin

Differential Revision: D7302454

fbshipit-source-id: da066febccebf0db54fffe00eac06f9ea68bd518