Track data-plane statistics for each AggregatePort
Summary:
This extends each BcmTrunk with a BcmTrunkStats object. An instance of
BcmTrunkStats tracks data-plane statistics for its corresponding trunk. It
accomplishes this by aggregating statistics for each member port of the trunk to
which it belongs.

To ensure these statistics are precise, BcmTrunkStats accumulates statistics
only for ports which are both

-configured as members of the trunk, AND
-programmed as members of the trunk

Depends on D7319964.

Reviewed By: boryas

Differential Revision: D7300664

fbshipit-source-id: e524fdf3a93633cf4a081e6a594323df0de0c0c2