Make BcmPort::portStats_ thread-safe
Summary:
BcmPort::portStats_ is read from and written to in the
updateStatsThread and may be written to in the updateThread (via
BcmPort::updateName()). This guards BcmPort::portStats_ with a read-write lock
to render such concurrent access safe.

This further factors out BcmPort::portStats_, in addition to the time when
portStats_ was last dervied, into a wrapper class BcmPortStats.

At the present time, BcmPortStats only serves as a convenience but in the
future, I plan on extracting all statistics functionality out of BcmPort and
into BcmPortStats.

Reviewed By: boryas

Differential Revision: D7320003

fbshipit-source-id: e7f53ed5847f5619d80d0ab41356cfe40cc0f5bf