use SwitchState when determining if safe to update interface mode
Summary:
We have some complex legacy logic for when we should avoid
updating the interface mode of a port during program(). This used to
block changing the interface mode if the port is up. To determine if
port is up, we actually ask the ASIC.

However, this actually introduces a race condition where a port could
flap quickly and by the time we get to program() the port is back up
in hw. Relying on the switch state should still be down if we are
programming the state update that resulted from downing the port.

We used to need to do this because there was a period during startup
where the operstate of a port was not accurate and relying on it led
to port flaps. However, I fixed that last half as part of the setup
for adding QsfpCache, so it should be safe to rely on switch state
instead of asking the asic, giving us a window to properly program
port settings during a flap.

Long term we should really try to get rid of this logic altogether and
always program the interface mode if it changes. Will require a bit
more testing though and this is smallest change to do for now to stem
the issues from KR ports showing up in prod.

Differential Revision: D7363218

fbshipit-source-id: 4e0309e7f641d119a2aaa06b1b822525fdf2ce42