Decouple core LACP logic from SwSwitch
Summary:
Previously, objects defined in LacpMachines.h called into SwSwitch to
transmit LACP frames and to update the SwitchState.

To faciliate unit-testing, a layer of indirection between the primary LACP
implementation and SwSwitch is introduced in the form of the abstract class
LacpServicerIf. This class represents the interface that must be fulfilled by
any object which can provide the services required by the core LACP logic (ie.
LACP frame transmission and specific SwitchState updates).

In effect, all instances of SwSwitch* are replaced with LacpServicerIf* in
LacpController.{h,cpp} and LacpMachines.{h,cpp}.

Reviewed By: boryas

Differential Revision: D7402567

fbshipit-source-id: 8d64e19b68a6251f61b151bb2207271f0f73c761