Test LACP can handle receiving a frame on a down port
Summary:
This adds a test for D6184510, which fixed a crash observed early on in
the production deployment of LACP.

At the time, neither reproducing the crash nor confirming the crash had been
fixed was possible as doing so required an LACP frame reception event occuring
close enough in time to a port down event so as to reorder the two events from
the perspective of the agent.

With the recent modifications, it is now possible to not only faithfully
recreate the crash but also test the correctness of the bug fix, demonstrating
that the LACP implementation organization allows for fine-grained testing of the
entire protocol stack.

[1] Corresponding to efe9853a8eff0dedee0242a4658ec365e3f5a345 on Github.

Depends on D7481701.

Reviewed By: boryas

Differential Revision: D7482587

fbshipit-source-id: cf2a05449f9661a82345eaf9f31e45525a90760e