Add synchronization primitive for ArpTest.sendRequest
Summary:
ArpTest.SendRequest has been disabled because it's been correctly
identified as being flaky by TestWarden. It fails under certain builds because
of asynchronous ARP expiration.

For instance the invocation of

  testSendArpRequest(sw, vlanID, senderIP, targetIP);

installs an ARP entry which, if it expires in time, will render

  EXPECT_HW_CALL(sw, stateChangedMock(_)).Times(1);

false.

This implements a new synchronization primitive- WaitForSwitchState- that will
allow for correctly handling the asynchrony that has led to ArpTest.SendRequest
being flaky.

Reviewed By: aeckert

Differential Revision: D7529101

fbshipit-source-id: 43006ce9769b1567954c7bdc60b8f17c9b6fa9f5