dont unselect after every module selection
Summary:
This should not be needed and slows down scanning all ports
because we always reset muxes to zero after accessing a port. If we
have a single layer this only adds a single unnecessary write (or two
if we select the same port twice). However, for multiple layers, it
will add (2n - 1) extra writes, where n is the number of muxes we need
to properly configure (or just 2n if we try to select the same port
two times in a row).

This seems like unnecessary overhead and both old and new pca9548
classes are smart enough to figure out what to select if something
else is already selected.

Reviewed By: boryas

Differential Revision: D7154143

fbshipit-source-id: 22e16cc46f4f866ad778334bf0f79b4193190507