add unit testing for PCA9548MuxedBus
Summary:
This changes the CP2112 class to derive from a mockable
virtual interface and then adds some tests for the module selection
logic added in D7154145.

Keeping this separate as courtesy to reviewers/because the interface
changes to cp2112 are a bit intrusive and would like feedback on that
independent of original diff.

Differential Revision: D7171975

fbshipit-source-id: c9b56d64d4edc1c161d09bc774e1059f87ca8908