Initial COPP system test
Summary:
Send a packet that looks like BGP and make sure it increments the
highpri cpu counter.  Then send a packet to the switch's router interface (not on BGP's port) to similarly test the midpri queue.

Identical test for Facebook only that uses Broadcom's packet_trace call.  Note that this does not work on wedge40.

Differential Revision: D7253949

fbshipit-source-id: f868285b6fbd3811b389fbbce81c1bb522015318