Fix QSFP tests
Summary:
Two QSFP (setRateSelect and setCdr) tests were failing because of QSFP changes (in setCdrIfSupported) that suppressed
calls to writeTransceiver when not needed.

There was also a new routine that was timing dependent
(ensureTxEnabled).  This was replaced with a MOCK routine for testing.  This
explains the intermittent failures of the setRateSelect QSFP test.

Differential Revision: D7649958

fbshipit-source-id: 839cdadcbc442e223085965a2686e38b4be2c0ee