Add a featuresDesired flag to selectively ask for (or not) features
Summary:
During bcm_tests we don't want regular linkscan, packet events
interfering with tests. So add a parameter to selectively ask for these
to be disabled. This diff is a noop as far as tests and production are
concerned. Subsequent diffs will ask for not setting up linkscan and
packet rx for tests.

Differential Revision: D7732249

fbshipit-source-id: 358373661ae25b5197a953dfaed6e8b4037016d4