ensure cache is dirty if we unsuccessfully read a page
Summary:
This is root cause for issues w/ ports coming up in KR mode
across reboots, which has been hitting us pretty consistently in
disruptive upgarades lately. More detail in the task, but the basic
problem is that it is possible to have an error reading from the
transceiver s.t we skip reading DOM data in to one of our cached pages
(in this case page0), but then treat the unset data as valid and
respond to thrift requests anyways.

In this case, page0 is all 0 data as we had an error populating it
through WedgeQsfp::readTransceiver(). Then later, when agent asks for
what type of cable is plugged in, we look for the DEVICE_TECHNOLOGY
field in this zeroed page, yielding (naturally) 0. A 0 DEVICE_TECH
field indicates optic in our current logic, which leads to agent
trying to program the wrong interface mode and causing this issue.

Long term, I think let's remove any dynamism from this determination
and just set interface mode in config, but this should fix stuff for
now.

Differential Revision: D7818515

fbshipit-source-id: 13eb354f812251e104474960ad5ce149f8cda589