skip customizing on missing/copper ports
Summary:
Was double checking behavior once I changed reads/writes over
I2C to raise an UsbError in D7818515 and noticed that we are doing a
lot of work trying to customize ports which will never
succeed. Furthermore, because we now raise in the middle of
customizeTransceiver for these ports, we never reach the end of that
function and set needsCustomize_ back to false.

This means that we try to customize missing or copper ports every time
refresh is called for that transceiver, which is a lot of wasted work
and creates logs likely to cause alarm upon investigation. This
changes the behavior to only try to customize ports with something
plugged in that is not known to be a copper cable.

Differential Revision: D7826302

fbshipit-source-id: b814b8a04c7ce7df4fa27720ce8a0f3f36b31cfa