make nextHopAddrs thrift default instead of required
Summary:
In order to support UCMP, next hops will have to have
associated weights. Currently, next hops reside in a required list of
binary addresses. We want to evolve this to be a list of "NextHopThrift"
structs. To start, we will downgrade nexthops from required to default.
Since thrift default objects do get written out with default values, this
should be safe for clients that still view it as required. Though we should
do some double writing to be safe

Reviewed By: tfangit

Differential Revision: D7538055

fbshipit-source-id: f6dc0c751a412cc7d88bda7c2941298b581bad4e