Summary: Change the way how qsfp_service determines how many lanes are in use
In wedge_agent there are 4 software port objects created for each physical port,
one per each lane. So the QsfpModule determines how many SW ports
are associated with current transceiver to get the number of lanes
and determine if customization is needed. Change it to count channels
that are used from SW ports.

Signed-off-by: Stepan Blyschak <stepanb@mellanox.com>