Support multiple LacpControllers in LacpServiceInterceptor
Summary:
Previously, LacpServiceInterceptor supported a single LacpController
LacpController. This extends LacpServiceInterceptor to allow it to support any
number of LacpControllers.

Support for multiple LacpControllers in LacpServiceInterceptor is necessary for
implementing LacpTests that test link aggregates with more than one link in
them.

This further adds a non-trivial destructor to LacpServiceInterceptor to avoid
failing a CHECK in folly::~AsyncTimeout(). Specifically, folly::~AsyncTimeout()
invokes EventBase::cancelTimeout(AsyncTimeout*), which requires that it be
executing in the thread belonging to the AsyncTimeout's EventBase.

Without this, tests in the LacpTest suite can fail non-deterministically during
LacpController destruction, and
LacpTest.nonAggregatablePortTransmitsIndividualBit does so reliably.

Reviewed By: boryas

Differential Revision: D7725681

fbshipit-source-id: 127292c90692d5495bb01727593fff36c25c06c6