COPP tests were timing out - increase wait time to account for
Summary:
When we send a packet through a front panel port, there are a bunch
of sources of delay before the corresponding packet counter is incremented.
There are:
  - internal SDK statistics thread  - runs once per second
  - internal agent statistics thread - runs once per second
  - any thread scheduling processing delays

The original test assumed that the thread scheduling delays were negligible,
but the test was still occasionally failing.  Add 0.5 seconds to account for
this time.

Adding more time to sleeps sounds like a bad thing, but I believe we've
characterized the sources of delay enough here...  perhaps we rewrite
as a series of retry tests if this works.

Differential Revision: D7897507

fbshipit-source-id: 8084c5681075e398cfa1257c3cec3684f2651ab4