replace RouteNextHop with NextHop, ResolvedNextHop, and UnresolvedNextHop
Summary:
RouteNextHops is somewhat awkward, so before adding weights to this
part of the system, I'd like to refactor it a bit.

My two complaints with the current design are
1) RouteNextHop is used in a place where it should not be (BcmHostKey)
2) RouteNextHop does not allow for more specific types and has a bunch of
   confusing, redundant constructors.

While this diff does not fully realize the goal of using the compiler to help
us ensure we only pass legit ResolvedNextHops to hw (or store them in fwd
period), it is a good first step. I think it also reduces some of the
confusion around Resolved vs. Unresolved NextHops.

Reviewed By: SamanKazemkhani

Differential Revision: D7538406

fbshipit-source-id: 4908cc8d52b95f0dc67fe57cf1de12f68b2565af