Work around BRCM bug which hits 6.4.10 for new FP groups post warmboot
Summary:
Throw in the towel and leak BRCM implementation details in qset
comparison. Bug is fixed in 6.5.12+ but backporting the patch is risky
per BRCM. So just ignore the extra qualifier added by SDK code.
Add a bcm_test to ensure that if assumptions ever change we will
get a failure.

We could go back to doing the old way post 6.5.12.

At any rate, this should unblock some of the QoS work we want to
do. But we will have to be careful about adding any new FP groups
on 6.4.10.

To truly guard against any new group showing up or QSETs changing,
go back to initing the FP if either of these conditions occur.

TODO: Add another test to check that only 2 FP groups exist post
init. This needs a little bit of a tweak in FakeSdk, so breaking
into 2 diffs.

Reviewed By: ninas

Differential Revision: D7448521

fbshipit-source-id: 19f49eef58fb419f54507b124378acf94d2951ca