Add a test for ensuring that FP groups are in desired state post init
Summary:
As title. The only new thing here is that I made
test a friend rather than making the function to be tested public.
It seems a worthwhile trade off for maintaining encapsulation
while still running tests on private methods.

ps: I did choose the opposite approach for static linkScanCallback
api in D7502678, since
a) It was static and does not have access to object data
b) By using it as a callback we were exposing it to external code anyways.

Differential Revision: D7621618

fbshipit-source-id: f04c5cf0a65bac2d5bcda2afa7b1905bf1bd774d