Cleanup our own static variables properly
Summary:
When unit is destroyed, set bcmInitialized to false.
In wedge_agent it doesnt quite matter that we dont uninit this
properly since
a) We don't actually destroy the process and just _exit the
process
b) On process restart bcmInitialized static variable is inited
to false and we do proper initialization.

However if you want to do warm boots w/o exiting process (ala
bcm_test warmboots) we need to cleanup properly.

Reviewed By: plapukhov

Differential Revision: D7986672

fbshipit-source-id: 52c0b4c6051e2742410527b733ab21a4253a44a7