XLOG
Summary:
We've run into a few outages where the blocking behavior of GLOG combined with it having a global logging lock has resulted in breaking forwarding due to an over-utilized disk, or even just logging too heavily. XLOG with async=true moves log flushing to its own thread which drops after filling a buffer, which should be much safer for us.

To accomplish this, I ran:
a crummy hand-made python script to insert:
  #include <folly/logging/xlog.h>
at the end of every #include block in a file that uses logging.
Then ran:
  find ./fboss/agent/ -type f -exec sed -i 's/VLOG(/XLOG(DBG/g' {} \;
  find ./fboss/agent/ -type f -exec sed -i 's/VLOG_IF(/XLOG_IF(DBG/g' {} \;
  find ./fboss/agent/ -type f -exec sed -i 's/VLOG_IS_ON(/XLOG_IS_ON(DBG/g' {} \;
  find ./fboss/agent/ -type f -exec sed -i 's/\bLOG(/XLOG(/g' {} \;
  find ./fboss/agent/ -type f -exec sed -i 's/XLOG(ERROR/XLOG(ERR/g' {} \;
To substitute XLOG for VLOG. We have one callsite of VLOG_EVERY_MS which has
no out of the box 1-1 substitute so I will fix that in a follow up diff.

The --logging flag is set in wedge_agent.sh: currently it is the
equivalent of '-v 2' so I need to clean that up a bit.

In a follow-up diff I intend to focus on "scouring GLOG from our codebase and binary" though this represents 95%+ of that effort.

Reviewed By: simpkins

Differential Revision: D8001334

fbshipit-source-id: 761465c80145faa9f6c818e3acc1a98d1badcfe2