Red diff unused code in Bcm{Switch, Unit}
Summary:
Post D7888008, we no longer have the notion of separate unit vs BcmSwitch
initializations. So get rid of the extra BcmSwitch constructor.
BcmUnit::attach() was interesting, in that it was just calling
BcmUnit::attach(std::string) and just using a empty string arg.
BcmWarmBootHelper was then using this empty string as a means of
determining whether it should support warm boot or not.
This is non unintuitive and unnecessary. It seems the distinction
was being made for bcm_test vs wedge_agent. However until we actually
write the warm boot files tests won't do warm boot either. So
no need to have 2 APIs

Reviewed By: boryas

Differential Revision: D7996069

fbshipit-source-id: 0dda46dc61bd610cef37db0adceb724bc9b4ba0f