Only expect a single port per tcvr on minipack
Summary:
Previous implementation always expected the agent to send
over all 4 ports that make up a transceiver. This caused minipack to
always think it was missing information from the agent and never
actually customize ports.

This adds an argument to the QsfpModule constructor to pass in how
many ports we expect to be part of this transceiver.

Differential Revision: D8060936

fbshipit-source-id: f5d64abcf1c94b72653bbfdb7b93590dfc64d8ba