add next hop weights to SwitchState
Summary:
This diff introduces weights to the NextHop objects.
The majority of the complexity lies around recursive resolution
with the possibility of weights. We originally tried ECMP is weight 1,
but that has some incorrect behavior in weird cases like two recursive
next hops both having the same resolved next hop in the next hop set.

There is no official, widely accepted standard for how this should be
done, so I went for an option that has desirable behavior
under a full UCMP configuration but reverts back to ECMP if any next
hop at any level in a recursive route resolution tree is ECMP. The
approach uses ECMP next hop <==> weight 0. This has the advantage of
slotting in freely into LCM, GCD calculations in the algorithm with
natural results and minimal "if" conditions. Naturally, this has the con
of occasionally obstructing the logic, so I attempted to thoroughly document it.

If people prefer a simpler, more explicit approach, I'm happy to change this.
A few alternatives I considered were:
1) forbid weights from thrift when the next hop is not resolved
   This is the simplest in recursive resolution, but it is unclear
   if it easily/cleanly do-able. I can also imagine situations where
   the route table changes later and we get surprised in a bad way
2) Flatten recursive weights in a simpler way. Something like
   default to ECMP with any multihops in recursive resolution, or
   use the top-level weight of the nh for all the recursively
   resolved next hops.

Reviewed By: tfangit

Differential Revision: D7681134

fbshipit-source-id: a3574ce203f1d72c5177af25d276e82bfafcab0a