Move logic for dumping and retrieving warm boot state to warm boot helper
Summary:
As titled. Logic for dumping warm boot state was in BcmUnit,
while that for retrieving state was in BcmWarmBootCache. Instead
move it to warm boot helper. This is better encapsulation in itself,
but is necessary prerequisite for introducing new warm boot helpers
(disc backed, ram backed). Since now we can't assume that there is
always a single way to dump/retrieve warm boot state

Reviewed By: plapukhov

Differential Revision: D8036633

fbshipit-source-id: 48bad8a14326ef99dbb6d83ce055684d10c39716