remove racy check in dbg mode in NeighborCache
Summary:
We have a DCHECK at the top of the flushEntryFromSwitchState
function that checks that the entry we are flushing from the switch
state is not in the cache. This gets called as part of the state
update when we are removing/expiring entries. This is problematic for
a number of reasons:

1. not thread safe.
2. not necessarily true.

The fact that it was a DCHECK caused this bug to stick around for
years, but we had a dbg build canaried on minipack and hit this.

Differential Revision: D8207300

fbshipit-source-id: 82f119d81906494cd830af92e19c24e1db34e69e