Refactor Bcm Cosq Manager: fix counters bug
Summary:
The main goal of this diff is to abstract BcmCosQueueManager, which is the parent class of BcmPortQueueManager and BcmControlPlaneQueueManager.
As we realized that there're some special handling for regular port queue and control plane queue separately. For example, only port queue supports
alpha and aqm, while only cpu queue supports Wedge40 platform. We should probably treat BcmPortQueueManager and BcmControlPlaneQueueManager differently.

Therefore, I introduced BcmCosQueueManager, which defines several important virtual functions for the child class to implement their own ways. Meanwhile,
we also keep most of the common features implementing in BcmCosQueueManager so that we can just reuse them.

After refactoring, we can have BcmControlPlaneQueueManager to handle both Wedge40 and Wedge100 queue stats counters. So this diff will be a final fix for S160512.

Differential Revision: D8208767

fbshipit-source-id: d24580840635ac89b1cf22457b7bd4f4cc74bd33