Enable portTable initPorts during initTables
Summary:
Currently portTable_ won't initPorts in BcmSwitch::initTables(), therefore, after initTables() BcmPortTable will be empty.

Besides, I also fix the old way to generate BcmTestPort cause in the old way, we first generate a unique_ptr<BcmTestPort> and then
we get the raw pointer out, but we also use std::move the unique_ptr to add it into the global map ports_. In this case, when we're
trying to return that raw pointer, the old unique_ptr might be deleted and leave this raw point be stale. The fix is no using any
temporary variable and avoid creating-and-then-deleting pointer.

Differential Revision: D8374482

fbshipit-source-id: ce0f08c46b2a6a1904dce70680ced9ba5eb856d6