refactor SwSwitch ucmp tests to be more compact
Summary:
Lots of copy paste in the tests previously. They all have a similar
structure and reuse lots of constants, so jam that into a text fixture.
New code isn't particularly elegant, but I think it is pretty compact and
makes it easy to define the next set of tests I'm working on.

Reviewed By: aeckert

Differential Revision: D8380217

fbshipit-source-id: 348c12f34e36d589af998c70401bc806da2463da