Add missing return statement
Summary:
This was missed in D8392803. However clang
is smart enough to see that we unconditionally assert
and there is no need for return. So everything was
green on the diff since we build tests in dev mode
using clang.
However if you build with opt mode with gcc, the
error shows up.

Differential Revision: D8458560

fbshipit-source-id: 9ad2b1c37e3d5e99ccbd5bbb4f74e818beed762a