Allow reset config for PortQueue
Summary:
Right now, when we try to update a port queue setting, we clone the original one and then update the value if the setting is not optional in config side.
It's OK for most of cases but will fail in the "reset" scenario.
Let's say, we program aqm to queue2 and suddenly we wanna remove the aqm setting for queue2. The old code will clone the current queue2 SW state and then
update the config settings. But since we want to remove the aqm setting, in config, the queue2.__isset.aqm is false. Therefore, the old code will never
remove the aqm setting.
The fix is easy: just create a new PortQueue object and then update the attributes based on cfg::PortQueue. Thus, we can get the correct Queue state to
program to H/W.

Differential Revision: D8388320

fbshipit-source-id: fdc938cf38d3700813b651e3a2bcd187c8751559