Refactor aqm to a list struct in PortQueue
Summary:
Current AQM struct can't support config different thresholds for early_drop and ecn at the same time. To enable that, we need to change
the data structure to a list so that we can add more behavior types in the future(considering BRCM can support different curves).
Also I've add a check in ApplyThriftConfig to forbid to have the same behavior AQM in aqms list for the same queue.

And there're no users using AQM in prod, so it'll be a good time to refactor it.

Reviewed By: boryas

Differential Revision: D7772964

fbshipit-source-id: b69a216d9071575306edf1a33cc25260c0027cdd