Add exception throwing lookup APIs in BcmAclTable
Summary:
Similar to pattern elsewhere. Add get*() apis
in addition to get*If() apis. These can be used in
places where we are guaranteed that lookup will succeed.
In those places, using the exception APIs conveys are
intention more clearly and also silences infer warnings
like those in D8383681. The latter arise because we
use FOLLY_NULLABLE as a attribute to the If lookup
APIs in BcmAclTable.

Reviewed By: schischi

Differential Revision: D8548065

fbshipit-source-id: f3983c6b7aaf13a9f4216a13898947dd9463c191