Add BcmRtag7LoadBalancer and BcmRtag7Module objects
Summary:
RTAG7 is a hashing engine common to both the Trident2 and Tomahawk
switching ASICs. The engine is comprised of two independently programmable
modules (refered to as module A and module B) with 2 hash outputs per module.

The BcmRtag7LoadBalancer singleton owns the entirety of the RTAG7 hashing
engine while each module is managed by a separate BcmRtag7Module object.

Because LoadBalancerMap can represent an abitrary hash-based load-balancing
scheme, it is the responsibility of BcmRtag7LoadBalancer to schedule the RTAG7
engine so as to fulfill the requirements of all LoadBalancers (or determine that
the configuration is impermissible).

Because the current configuration only allows up to two load-balancers (one for
ECMP and one for LAG), and RTAG7 exposes two independent hashing modules, all
LoadBalancer configurations are trivially permissible by dedicating a module to
each LoadBalancer.

Differential Revision: D7987065

fbshipit-source-id: c6ea03e50926c29d2f0c16f3e781cec622170942