Only cos queues under the cosQ manager
Summary:
Ports for chips that support the higig interface (which we don't use)
have 2 exta queues besides the cosQs.  When calling
getCurrentQueueSettings there is a call to getAqms which calls
bcm_cosq_qport_discard_get which (as of 6.5.13) checks to see if the
resulting queue is a cosQ and errors out if not.  So we now check the
queue of the gport before adding to our cosQ manager to make sure that
the queue number is less than the number of cosQs on the device.

Also, instead of pushing gports onto the unicast and multicast queues
they are now ordered by their assigned queue number in hardware.  That
means that out queueIdx is the same and the actual queue number and we
do not have any entries for a queueIdc > number of cosqs on this device.

Differential Revision: D8414410

fbshipit-source-id: 159b529df4000c0c634d7d0661631629de1eb6b5