Support warm-boot for LoadBalancers
Summary:
This implements warm-boot support for LoadBalancers.

It accomplishes this by extending BcmWarmBootCache to populate the state of the
RTAG7 load-balancing engine and using this state to avoid programming the engine
if unnecessary.

Specifically, any LoadBalancers in the SwitchState derived from the warm-boot
file will be handled in BcmRtag7Module::init(...). In this method, the
pre-existing state of the module (captured by the newly introduced
BcmRtag7Module::ModuleState) is retreived from the warm-boot cache and used to
skip over hardware programming.

Differential Revision: D8344382

fbshipit-source-id: c2940ae6fa1cf2e6429fea976e3af9bc3cc095ce