Substitute (newly!?) racy NDP test for RA test
Summary:
Previous logic was: wait exactly 5 seconds for a NDP packet
to be sourced from the switch.  This was super stable, but something
changed (!?? - tracking as T31145754) and now this test is failing
sometimes because seemingly the packets are not always sent for all
hosts.

Added a RA discovery test in the mean time to test the jist of this
functionality in the mean time.  It's more stable and doesn't leave
a big whole in our test coverage.

Reviewed By: linhbui, jhshi

Differential Revision: D8680113

fbshipit-source-id: 9d7fa2243bae417c2b32893fe5f5080778af1684