Expose FBOSS ECN marking counters in ODS
Summary:
T25421873 introduces support for ECN in FBOSS. This patch exposes ECN
marking counters.

The counters are not available via opensl_* API. Furthermore, the counters
are only available on 6.5.13 or later via bcm_stat_get (refer Broadcom
case CS5166366). The counters are exposed as monotonic counters via
Thrift API.

Note that with this patch, ECN counters will appear in ODS for 6.4.10 as
well, but those counters will always be 0. Would like to hear thoughts
on that.

Reviewed By: boryas

Differential Revision: D8673832

fbshipit-source-id: d0cc17eff276570d1996addb0bdd3c173159052a