Enable applying cpu cosq setting from config
Summary:
This diff will apply cpu cosq settings from config.
I've made some effort to reuse the logic of applying non-cpu port queue setting into ThriftConfigApplier::updateControlPlane() so that
in the future, if we need add a new attribute or change an existing one, it'll be capable for both cpu queue and regular port queue.
The main logic to program to BRCM has been implemented in the previous diffs, so this diff focus on how to integrate S/W and H/W to finally
support applying cpu cosq settings from config.

1) add some state tests in the future diff to make sure we won't apply the same config again
2) remove FbBcmCosManager::configureRxCosq() and configureCPUCosqLimits() in the next diff

Differential Revision: D8098570

fbshipit-source-id: 9e4231fb115fceb3257f8770c3fd2aafd86dfe70